Convert WaitingForDriver modal to a page
