+++ client/src/Client/hooks/useNavigateUser.ts
@@ -37,11 +37,11 @@ const useNavigateUser = () => {
     }
     if (actualUser?.role === UserRoleEnum.Requester) {
       if (activeRide) {
-        if (
-          activeRide.state === RideStateEnum.WaitingForDriver ||
-          activeRide.state === RideStateEnum.Canceled
-        ) {
+        if (activeRide.state === RideStateEnum.Canceled) {
           return navigate('/passenger/order-ride');
+        }
+        if (activeRide.state === RideStateEnum.WaitingForDriver) {
+          return navigate('/passenger/searching-driver');
         }
         if (activeRide.state === RideStateEnum.DriverArrived) {
           return navigate('/passenger/driver-arrived');
